Adds an easier way to find the base turf of a tile.
A list of indices of equal length to the number of tiles is created. The value of the cell is equal to the tileId of its baseTurf.

For example if you had space/0, lattice/1, plating/2 underplating/3, floor_steel/4, floor_hydro/5 the array would be

0 0 1 2 3 3

The idea here is to stick only one base_turf per tile, the only one that has more than one base_turf is wood anyway